Move files into views/ and endpoints/ directories
