+++ src/sentry/integrations/slack/endpoints/command.py
@@ -5,10 +5,10 @@ from django.http import Http404, HttpResponse
 from rest_framework.request import Request
 
 from sentry.api.base import Endpoint
-from sentry.integrations.slack.link_team import build_linking_url
 from sentry.integrations.slack.message_builder.help import SlackHelpMessageBuilder
 from sentry.integrations.slack.requests.base import SlackRequestError
 from sentry.integrations.slack.requests.command import SlackCommandRequest
+from sentry.integrations.slack.views.link_team import build_linking_url
 from sentry.models import Identity, IdentityProvider
 
 logger = logging.getLogger("sentry.integrations.slack")
